Refac: refactor model modules
