+++ src/careamics/models/model_utils/__init__.py
@@ -1,0 +1,5 @@
+"""Model utils."""
+
+from .activation import get_activation
+
+__all__ = ["get_activation"]
