+++ src/careamics/models/unet/__init__.py
@@ -1,0 +1,5 @@
+"""UNet architecture implementation."""
+
+from .unet import UNet
+
+__all__ = ["UNet"]
